// src/global/enums/users.ts
// when changing or adding any role then change
/* //? add a user model in 1. (save) pre method in add if condition , 2.isUserFindMethod in add aggregation,
 */
export enum ENUM_USER_ROLE {
  superAdmin = "superAdmin",
  admin = "admin",

  generalUser = "generalUser",
}
export enum ENUM_GENDER {
  MALE = "male",
  FEMALE = "female",
  OTHER = "other", // other
}